QuizNum Long 타입  및 제약 조건 변경

+++ src/main/java/com/example/demo/jpa/QuizRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.jpa;
 import com.example.demo.entity.QuizEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizRepository extends JpaRepository<QuizEntity, Integer> {
+public interface QuizRepository extends JpaRepository<QuizEntity, Long> {
 
 }
